backend/requirements.txt
fastapi
pydantic
requests
google-generativeai
google-cloud-aiplatform
python-dotenv
google-adk